Add mixin class to allow function reuse across parser
